backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.4.2
python-multipart==0.0.20
openai==1.62.0
aiosqlite==0.21.0
pyarrow==21.0.0

